rag/retrieval/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
boto3>=1.34.0
openai>=1.0.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-openai>=0.0.5
python-multipart>=0.0.6  # For file uploads if needed
python-dotenv>=1.0.0
pickle5>=0.0.11  # For compatibility if needed